scripts/requirements.txt
notion-client==2.2.0
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
Pillow==10.0.0
beautifulsoup4==4.12.2
markdownify==0.11.6
